Added different URI paths for different methods
Added new URI paths for different fetch methods. If a command is requested over DNS (like how the CF worker will fetch it) there's no HMAC attached. If a command is requested over HTTP, it is assumed that the implant is requesting it so an HMAC is returned along with it